update code for neuromodulated masking. neuromod net knowledge is preserved using scp. implemented code to generate sparse mask from the neuromod net using kwinner activation
 On branch seq_tasks
 Your branch is up-to-date with 'origin/seq_tasks'.

 Changes to be committed:
	modified:   cl_examples_sparsity.py
	modified:   deep_rl/agent/PPO_agent.py
	modified:   deep_rl/component/task.py
	modified:   deep_rl/network/network_bodies.py
	modified:   deep_rl/network/nm_net.py

+++ deep_rl/component/task.py
@@ -154,7 +154,9 @@ class CTgraph(BaseTask):
         else:
             self.state_dim = env.observation_space.shape
 
+        # for labels randomly sampled from a uniform distribution
         self.task_label_dim = 64
+        # for one-hot labels
         #self.task_label_dim = 4
         self.env = self.set_monitor(env, log_dir)
 
@@ -190,10 +192,12 @@ class CTgraph(BaseTask):
 
     def get_all_tasks(self, requires_task_label=False):
         if requires_task_label:
-            #tasks_label=np.random.uniform(low=0.,high=1.,size=(len(self.tasks),self.task_label_dim))
-            #tasks_label = tasks_label.astype(np.float32)
+            # one-hot labels
             #tasks_label = np.eye(len(self.tasks)).astype(np.float32)
+            #tasks_label[tasks_label == 0.] = -1.
+            # randomly sampled labels from uniform distribution
             tasks_label=np.random.uniform(low=-1.,high=1.,size=(len(self.tasks),self.task_label_dim))
+            tasks_label = tasks_label.astype(np.float32)
             tasks = copy.deepcopy(self.tasks)
             for task, label in zip(tasks, tasks_label):
                 task['task_label'] = label
@@ -205,8 +209,12 @@ class CTgraph(BaseTask):
         tasks_idx = np.random.randint(low=0, high=len(self.tasks), size=(num_tasks,))
         if requires_task_label:
             all_tasks = copy.deepcopy(self.tasks)
+            # one-hot labels
             #tasks_label = np.eye(len(all_tasks)).astype(np.float32)
+            #tasks_label[tasks_label == 0.] = -1.
+            # randomly sampled labels from uniform distribution
             tasks_label=np.random.uniform(low=-1.,high=1.,size=(len(all_tasks),self.task_label_dim))
+            tasks_label = tasks_label.astype(np.float32)
             tasks = []
             for idx in tasks_idx:
                 task = all_tasks[idx]
